# fixes range deriving for constructs incl classes/interfaces