--- a/semanticdb-javac/src/main/java/com/sourcegraph/semanticdb_javac/RangeFinder.java
+++ b/semanticdb-javac/src/main/java/com/sourcegraph/semanticdb_javac/RangeFinder.java
@@ -16,10 +16,13 @@ import java.util.regex.Pattern;
 import static com.sourcegraph.semanticdb_javac.Debugging.pprint;
 
 public class RangeFinder {
-  private static final HashMap<CompilationUnitTree, String> contentsCache = new HashMap<>();
-
   public static Optional<Semanticdb.Range> findRange(
-      TreePath path, Trees trees, CompilationUnitTree root, Element element, int startPos) {
+      TreePath path,
+      Trees trees,
+      CompilationUnitTree root,
+      Element element,
+      int startPos,
+      String source) {
     LineMap lineMap = root.getLineMap();
     Name name = element.getSimpleName();
     if (name.contentEquals("<init>")) name = element.getEnclosingElement().getSimpleName();
@@ -27,7 +30,7 @@ public class RangeFinder {
     int endPos = (int) trees.getSourcePositions().getEndPosition(root, path.getLeaf());
     // false for anonymous classes
     if (name.length() != 0) {
-      startPos = findNameIn(root, name, startPos);
+      startPos = findNameIn(name, startPos, source);
       endPos = startPos + name.length();
     }
 
@@ -45,21 +48,10 @@ public class RangeFinder {
     return Optional.of(range);
   }
 
-  private static int findNameIn(CompilationUnitTree root, CharSequence name, int start) {
-    String content =
-        contentsCache.computeIfAbsent(
-            root,
-            (s) -> {
-              try {
-                return root.getSourceFile().getCharContent(true).toString();
-              } catch (IOException e) {
-                return null;
-              }
-            });
+  private static int findNameIn(CharSequence name, int start, String source) {
+    if (source.equals("")) return -1;
 
-    if (content == null) return -1;
-
-    int offset = content.indexOf(" " + name, start);
+    int offset = source.indexOf(" " + name, start);
     if (offset > -1) {
       return offset + 1;
     }