--- a/semanticdb-javac/src/main/java/com/sourcegraph/semanticdb_javac/SemanticdbVisitor.java
+++ b/semanticdb-javac/src/main/java/com/sourcegraph/semanticdb_javac/SemanticdbVisitor.java
@@ -107,7 +107,7 @@ public class SemanticdbVisitor extends TreePathScanner<Void, Void> {
     if (node instanceof JCTree.JCMethodDecl) {
       JCTree.JCMethodDecl meth = (JCTree.JCMethodDecl) node;
       CompilerRange range = CompilerRange.FROM_POINT_TO_SYMBOL_NAME;
-      if (meth.params.isEmpty() && (meth.sym.flags() & Flags.GENERATEDCONSTR) != 0L) {
+      if ((meth.sym.flags() & Flags.GENERATEDCONSTR) != 0L) {
         range = CompilerRange.FROM_TEXT_SEARCH;
       }
       emitSymbolOccurrence(meth.sym, meth, Role.DEFINITION, range);