--- a/tests/snapshots/src/main/generated/com/airbnb/epoxy/AfterPropsSet.java
+++ b/tests/snapshots/src/main/generated/com/airbnb/epoxy/AfterPropsSet.java
@@ -39,6 +39,6 @@ import java.lang.annotation.Target;
 //         ^^^^^^^^^^^^^^^ reference java/lang/annotation/RetentionPolicy#
 //                         ^^^^^ reference java/lang/annotation/RetentionPolicy#CLASS.
 public @interface AfterPropsSet {
-//      ^^^^^^^^^^^^^ definition com/airbnb/epoxy/AfterPropsSet#
+//                ^^^^^^^^^^^^^ definition com/airbnb/epoxy/AfterPropsSet#
 }