--- a/tests/snapshots/src/main/generated/com/airbnb/epoxy/AutoModel.java
+++ b/tests/snapshots/src/main/generated/com/airbnb/epoxy/AutoModel.java
@@ -36,6 +36,6 @@ import java.lang.annotation.Target;
 //         ^^^^^^^^^^^^^^^ reference java/lang/annotation/RetentionPolicy#
 //                         ^^^^^ reference java/lang/annotation/RetentionPolicy#CLASS.
 public @interface AutoModel {
-//      ^^^^^^^^^ definition com/airbnb/epoxy/AutoModel#
+//                ^^^^^^^^^ definition com/airbnb/epoxy/AutoModel#
 
 }