--- a/tests/snapshots/src/main/generated/com/airbnb/epoxy/CallbackProp.java
+++ b/tests/snapshots/src/main/generated/com/airbnb/epoxy/CallbackProp.java
@@ -45,5 +45,5 @@ import java.lang.annotation.Target;
 //         ^^^^^^^^^^^^^^^ reference java/lang/annotation/RetentionPolicy#
 //                         ^^^^^ reference java/lang/annotation/RetentionPolicy#CLASS.
 public @interface CallbackProp {
-//      ^^^^^^^^^^^^ definition com/airbnb/epoxy/CallbackProp#
+//                ^^^^^^^^^^^^ definition com/airbnb/epoxy/CallbackProp#
 }