--- a/tests/snapshots/src/main/generated/com/airbnb/epoxy/ControllerHelperLookup.java
+++ b/tests/snapshots/src/main/generated/com/airbnb/epoxy/ControllerHelperLookup.java
@@ -30,8 +30,8 @@ import androidx.annotation.Nullable;
  * be returned.
  */
 class ControllerHelperLookup {
-//^^^^^^ definition com/airbnb/epoxy/ControllerHelperLookup#`<init>`().
-//^^^^^^^^^^^^^^^^^^^^^^ definition com/airbnb/epoxy/ControllerHelperLookup#
+//    ^^^^^^^^^^^^^^^^^^^^^^ definition com/airbnb/epoxy/ControllerHelperLookup#
+//    ^^^^^^^^^^^^^^^^^^^^^^ definition com/airbnb/epoxy/ControllerHelperLookup#`<init>`().
   private static final String GENERATED_HELPER_CLASS_SUFFIX = "_EpoxyHelper";
 //                     ^^^^^^ reference java/lang/String#
 //                            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ definition com/airbnb/epoxy/ControllerHelperLookup#GENERATED_HELPER_CLASS_SUFFIX.