Uses text search derived from previous com.sun.source lsif-java (Kotlin edition), which is originally derived from georgewfraser's java language server. Closes #109 

Of particular note is that the current changes cause definition occurence to [not be emitted](https://github.com/sourcegraph/lsif-java/pull/121/files#diff-1c43dc4d39a78f8bfeee7dbff8bd285a2cd2c711e57d8053586575921be9230cL325). The reference to this no longer existing definition occurrence still gets emitted, we can either follow up with a PR or fix it in this PR. 

I feel like moving parts of `SemanticdbVisitor.semanticdbRange(...)` to the `RangeFinder` class to keep more of the logic together, but I will revisit this tomorrow.