@Strum355 2021-03-11
This branch:
```
[info] Benchmark                           (lib)  Mode  Cnt     Score    Error  Units
[info] CompileBench.compile            bytebuddy    ss   10  1371.189 ± 46.755  ms/op
[info] CompileBench.compile                guava    ss   10  1860.100 ± 94.912  ms/op
[info] CompileBench.compileSemanticdb  bytebuddy    ss   10  1684.466 ± 32.033  ms/op
[info] CompileBench.compileSemanticdb      guava    ss   10  2292.896 ± 45.351  ms/op
```

Current main:
```
[info] Benchmark                           (lib)  Mode  Cnt     Score    Error  Units
[info] CompileBench.compile            bytebuddy    ss   10  1379.276 ± 37.184  ms/op
[info] CompileBench.compile                guava    ss   10  1870.692 ± 95.841  ms/op
[info] CompileBench.compileSemanticdb  bytebuddy    ss   10  1662.590 ± 39.720  ms/op
[info] CompileBench.compileSemanticdb      guava    ss   10  2282.368 ± 70.845  ms/op
```